Extract flow mixins from BoxmanManager, slim core to sessions/runtime (#85 item 29, part 3)
Final move-only slice: snapshot/storage flows, top-level
provision/up/down/deprovision/destroy flows, compose orchestration,
control verbs, netlab lifecycle, and inspection/misc handlers move
verbatim into mixin classes under src/boxman/manager_parts/
(SnapshotsMixin, FlowsMixin, ComposeMixin, ControlMixin, NetlabMixin,
MiscMixin). The remaining core (553 lines) keeps __init__, the
_sessions registry and session_for_*, runtime plumbing, the netlab
property, cache registration, _run_parallel, and the parallel-worker
module functions.

Only in-code adjustments: snapshot_collapse/storage_compact reach the
_collapse_one_vm/_compact_one_vm staticmethods via SnapshotsMixin
instead of BoxmanManager (same function objects; importing
boxman.manager from the mixin would be circular).

Tests: patch anchors rehomed with the moved code were repointed
(boxman.manager.subprocess/shutil/os.path.isdir ->
boxman.manager_parts.flows.*, boxman.manager.time.sleep ->
boxman.manager_parts.snapshots.time.sleep). No test logic changed.

Refs #85
